+++ duckdb-vortex/src/include/vortex_scan.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "duckdb/main/extension_util.hpp"
+
+
+namespace duckdb {
+
+void RegisterVortexScanFunction(DatabaseInstance &instance);
+
+}
