+++ vortex-ffi/src/lib.rs
@@ -10,21 +10,19 @@ mod file;
 mod log;
 mod stream;
 
-use std::cell::LazyCell;
 use std::ffi::{CStr, c_char, c_int};
+use std::sync::LazyLock;
 
 use tokio::runtime::{Builder, Runtime};
 use vortex::error::VortexExpect;
 
-thread_local! {
-    static RUNTIME: LazyCell<Runtime> = LazyCell::new(|| {
-        // Using a new_multi_thread runtime since a current local runtime has a deadlock.
-        Builder::new_multi_thread()
-            .enable_all()
-            .build()
-            .vortex_expect("building runtime")
-    });
-}
+static RUNTIME: LazyLock<Runtime> = LazyLock::new(|| {
+    // Using a new_multi_thread runtime since a current local runtime has a deadlock.
+    Builder::new_multi_thread()
+        .enable_all()
+        .build()
+        .vortex_expect("building runtime")
+});
 
 pub(crate) unsafe fn to_string(ptr: *const c_char) -> String {
     let c_str = CStr::from_ptr(ptr);
