+++ vortex-ffi/src/stream.rs
@@ -46,7 +46,7 @@ pub unsafe extern "C" fn FFIArrayStream_next(stream: *mut FFIArrayStream) -> boo
 
     // TODO(joe): replace with futures::executor::block_on, currently vortex-file has a hidden
     // tokio dep
-    let element = RUNTIME.with(|runtime| runtime.block_on(async { inner.stream.next().await }));
+    let element = RUNTIME.block_on(async { inner.stream.next().await });
 
     if let Some(element) = element {
         let inner = element.vortex_expect("element");
